커뮤니티 사용자 조회(멀티건) 카운트 단위테스트

+++ src/main/java/egovframework/com/cop/com/service/impl/EgovUserInfManageDAO.java
@@ -67,7 +67,7 @@ public class EgovUserInfManageDAO extends EgovComAbstractDAO {
 	 * @throws Exception
 	 */
 	public int selectCmmntyUserListCnt(UserInfVO userVO) throws Exception {
-		return (Integer) selectOne("EgovUserInfManageDAO.selectCmmntyUserListCnt", userVO);
+		return selectOne("EgovUserInfManageDAO.selectCmmntyUserListCnt", userVO);
 	}
 
 	/**
